feat: worker pool is working now

+++ internal/service/source.go
@@ -1,6 +1,6 @@
 package service
 
 type Source struct {
-	Source string `json:"source"`
-	Rate   int    `json:"rate"`
+	Source string  `json:"source"`
+	Rate   float64 `json:"rate"`
 }
